Event module updates (#6)
* Docs skeleton

* Remove goto, fix registered players logic

While removing the GUI filters I inadvertantly broke the logic for
registered players. It would skip calling the handler if ANY of the
handler's conditional data did not have the player registered. Now
it properly checks all conditional events before deciding not to call
the handler.

* Formatting fixes

* Add support for matcher functions

This function will be run and the return value will determine if the
handler is called. I don't plan on ever using this in actual mod code,
it will only be used for the GUI module later on.

* Remove __version from global

This isn't actually needed since version changes will be kept track
of by the on_configuration_changed event itself.

* Remove needless (and broken) if statement

* Remove html from WIP docs

* Refactor dispatch_event and event.register

Dispatch_event how has a simplified structure to improve performance.
Additionally, the handler will be called for all matching conditional
events, and the conditional event name is passed in the event table.
This will make GUI filter matching possible. The function no longer
modifies the event's name, but instead gets the registry directly.
Also removed the newly added event matchers, since they are no
longer needed.

event.register has been slightly rewritten to improve performance.
The master handler is now registered in under the same conditions
as its registry is made, removing a redundant check.

* Improve handler reregistration, logging

Direclty re-register the handler in on_load by calling event.register,
instead of going through event.enable.
Log a warning if a conditional event's data was not found, to make
it clear that it should be handled by the modder in
on_configuration_changed.

* Remove unused parameter from shortcut functions

Bootstrap functions do not go through dispatch_event, and so do not
support the options table.

* Add usage examples for most functions

* Dynamically generate bootstrap event shortcuts

* Minor logging fix

* Add temporary class documentation

* Minor logging fix

* Documentation syntax fixes

* Minor syntax fixes