Change global structure, remove empty tables
Filters are now stored by event ID, then by player_index, instead
of the other way around. This is purely for consistency, as all
of a player's filters are kept together in the lookup tables.

A __size entry in each table keeps track of the table's size, and
when the table reaches zero, it is removed. This eliminates useless
tables being left in global, and is much more performant than using
table_size().