Gui module (#10)
* Import from RaiLuaLib, update paths

* Create handler_lookup table, better variable names

* First pass at GUI event filtering

* Remove unused parameter

* Fix documentation syntax

* Store filters with their event data

* Remove event module dependency, start refactor

* Separate functions to add templates and handlers

* Optimize generate_handler_lookup()

* Handler dispatching and filter returns

* Add register_events() function

* Documentation work

* Prototype helper function

* Fix error when registering custom-inputs

* Add remove_filters, allow init on config changed

* Return if a handler was dispatched

* WIP docs

* Remove redundant "train" from function name

* Store filters by handler in lookup table

This will let modders easily remove filters corresponding to certain
handlers or groups.

* Always manipulate filters by handler or group name

This makes the update_filters() function a bit slower and more bloated,
but SIGNIFICANTLY simplifies the management and usage of the
filters. The amount of data stored in global is unchanged, the extra
data is all kept in local lookup tables that are re-constructed during
bootstrap.

* Docs fix

* Change global structure, remove empty tables

Filters are now stored by event ID, then by player_index, instead
of the other way around. This is purely for consistency, as all
of a player's filters are kept together in the lookup tables.

A __size entry in each table keeps track of the table's size, and
when the table reaches zero, it is removed. This eliminates useless
tables being left in global, and is much more performant than using
table_size().

* Fix generate_filter_lookup()

* Minor consistency and docs fixes

* Minor docs fixes