New, significantly smaller translation module
This new module doesn't finagle with the output at all, it is left up to
the mod to deal with the translations. All it does now is spread the
workload over multiple ticks.

It also deals with a base-game bug where translation requests will
never be fulfilled if they're requested on the same tick that a save is
saved and reloaded.