Add gui.check_filter_validity()
This is a new REQUIRED setup function for the GUI module. This function
checks all GUI filters stored in global and if their corresponding
handler no longer exists (was removed between mod versions), removes
the filter. This is to protect against crashes when the GUI module would
try to call a handler that no longer exists.