Fix translations getting stuck sometimes
The code for parsing the locale identifier did not abort if the player
data already existed, causing the player to get added to the players
list in the language data multiple times. If the translating player
left, and had a duplicate in the players array, then they would get put
into the "waiting" state, but no other player would become the
translator. Hence, the translations would get permanently stuck.