Fix inconsistent crash in on_tick
Removing an entry from a table, then inserting into it, while iterating
with next() is unsafe. Lua might re-use the memory from the removed
entry for the new entry, which causes an error when next() tries to
iterate. To avoid this, insert the new entry, then delete at the end of
the loop.